Ensure Pybind Version Compatability (#21)
Ensure that pybind version is compatible

We find pybind from system by default. However, the default pybind
installed via apt for Ubuntu 20.04 and 22.04 is not compatible with the
version expected by jrl. This change adds a min required version so that
we dont find a system version that is incompatible with jrl.